pip-requirements.txt
python-binance==1.0.15
pandas==1.4.0
pyTelegramBotAPI==3.7.6
pytest==7.1.1
pytest-mock==3.7.0
coverage==6.3.2
